+++ Runtime/Decal/Gradation/SingleGradationDecal.cs
@@ -12,7 +12,7 @@ using net.rs64.TexTransCore.MultiLayerImageCanvas;
 namespace net.rs64.TexTransTool.Decal
 {
     [AddComponentMenu(TexTransBehavior.TTTName + "/" + MenuPath)]
-    public sealed class SingleGradationDecal : TexTransRuntimeBehavior, ICanBehaveAsLayer , ITexTransToolStableComponent
+    public sealed class SingleGradationDecal : TexTransRuntimeBehavior, ICanBehaveAsLayer, ITexTransToolStableComponent
     {
         internal const string ComponentName = "TTT SingleGradationDecal";
         internal const string MenuPath = ComponentName;
@@ -78,12 +78,31 @@ namespace net.rs64.TexTransTool.Decal
 
         private void OnDrawGizmosSelected()
         {
+            var preCol = Gizmos.color;
             Gizmos.color = UnityEngine.Color.black;
             Gizmos.matrix = transform.localToWorldMatrix;
 
             Gizmos.DrawLine(Vector3.zero, Vector3.up);
+
+            var step = 0.05f;
+            var halfStep = step * 0.5f;
+            var vecO = Vector3.zero;
+            var boxSize = new Vector3(halfStep, step, halfStep);
+            for (var i = 0.0f; 1.0f > i; i += step)
+            {
+                vecO.y = i + halfStep;
+                var gradCol = Gradient.Evaluate(vecO.y);
+                var bocGradCol = gradCol;
+                bocGradCol.a *= 0.5f;
+                Gizmos.color = bocGradCol;
+                Gizmos.DrawCube(vecO, boxSize);
+                Gizmos.color = gradCol;
+                Gizmos.DrawWireCube(vecO, boxSize);
+            }
             if (IslandSelector != null) { IslandSelector.OnDrawGizmosSelected(); }
+            Gizmos.color = preCol;
         }
+
         internal override IEnumerable<Renderer> ModificationTargetRenderers(IRendererTargeting rendererTargeting)
         {
             return DecalContextUtility.FilterDecalTarget(rendererTargeting, RendererSelector.GetSelectedOrIncludingAll(rendererTargeting, this, GetDRS, out var _), TargetPropertyName);