+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/ReinaS-64892/TexTransTool/compare/v0.10.7...HEAD)
 
+### Added
+
+- SingleGradationDecal の グラデーションの長さ が、設定値が 0 ~ 2 の間はスライダーになるようになり Transform から直接操作した場合に従来の描画になるようになりました (#1018)
+- SingleGradationDecal の ギズモ表示 が、グラデーションの設定された色を表示するようになりました (#1018)
+
+### Fixed
+
+- SingleGradationDecal の グラデーションの長さ が操作できない問題を修正しました (#1018)
+
+### Dependency
+
+- TTCE-Wgpu の オプション依存関係が v0.2.0 を要求するようになりました。(#1019)
+
 ## [v0.10.7](https://github.com/ReinaS-64892/TexTransTool/compare/v0.10.6...v0.10.7) - 2025-06-14
 
 ### Fixed